live36: pilot post-mortem fixes — temperature, loop guard, validator winsec
The first two live pilot episodes ran GREEDY and died in identical-action
loops (redis-t10068 killed by the guard at run 12; fluentd-3328 looped 531x
to the 2400 s drain cap). Three defects, three fixes:

- run_live36.sh now exports SWE_TEMP=0.6 explicitly: campaign.conf defaults
  to 0.0 ('stock/certified'), which contradicts the never-run-agents-greedy
  invariant and the census's verified 0.6 — the driver must not inherit it
- loop_guard: the action-block builder read until a blank line, and in this
  log format the block runs into the DEBUG/MODEL INPUT lines whose step
  numbers vary, so consecutive blocks never compared equal (fluentd was
  missed). The block now also ends at a DEBUG/INFO/MODEL INPUT/OBSERVATION
  marker; hardened rule detects runlen 90 on the banked failure log
- validate_glm_agents: winsec parsed with int() -> int(0.1)==0 -> division by
  zero on every 100 ms campaign; now float()

Both greedy episodes parked as run_1.greedy_bak (E7-flagged evidence, off the
DONE path so the re-run is clean).

+++ local_agents/kit/validate/validate_glm_agents.py
@@ -88,7 +88,8 @@ def roles_of(meta):
 def load_meta(rd):
     try:
         meta = json.load(open(f"{rd}/metadata.json"))
-        return meta, roles_of(meta), int(meta.get("winsec", 10))
+        # float, not int: int(0.1) == 0 and the 100 ms campaigns then divide by zero
+        return meta, roles_of(meta), float(meta.get("winsec", 10))
     except Exception:
         return {}, {}, 10
